refactor loader code
