+++ source/gui.c
@@ -15,6 +15,27 @@
 
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+    `rrc_gui_video_fix' uses code adapted from Brainslug:
+    Copyright (C) 2014, Alex Chadwick
+
+    Permission is hereby granted, free of charge, to any person obtaining a copy
+    of this software and associated documentation files (the "Software"), to deal
+    in the Software without restriction, including without limitation the rights
+    to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+    copies of the Software, and to permit persons to whom the Software is
+    furnished to do so, subject to the following conditions:
+
+    The above copyright notice and this permission notice shall be included in
+    all copies or substantial portions of the Software.
+
+    THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+    IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+    FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+    AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+    LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+    OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+    SOFTWARE.
 */
 
 #include <stdlib.h>
@@ -149,4 +170,63 @@ void _rrc_gui_set_mm_video_mode(int mode)
 GXRModeObj *rrc_gui_get_video_mode()
 {
     return VIDEO_GetPreferredMode(NULL);
+}
+
+/*
+    This code is part of Brainslug, adapted for this channel.
+    See copyright notice at the start of this file.
+
+    https://github.com/Chadderz121/brainslug-wii/blob/8ca49384452dcb7d41e90d002ba0f85b4e57bf57/src/apploader/apploader.c#L114
+*/
+void rrc_gui_video_fix(char region)
+{
+    /* Get video mode configuration */
+    bool progressive = (CONF_GetProgressiveScan() > 0) && VIDEO_HaveComponentCable();
+    bool PAL60 = CONF_GetEuRGB60() > 0;
+    u32 tvmode = CONF_GetVideo();
+
+    int r_rmode_reg = 0;
+    void *r_rmode = VIDEO_GetPreferredMode(0);
+
+    switch (tvmode)
+    {
+    case CONF_VIDEO_PAL:
+        r_rmode_reg = PAL60 ? VI_EURGB60 : VI_PAL;
+        r_rmode = progressive ? &TVEurgb60Hz480Prog : (PAL60 ? &TVEurgb60Hz480IntDf : &TVPal528IntDf);
+        break;
+
+    case CONF_VIDEO_MPAL:
+        r_rmode_reg = VI_MPAL;
+        r_rmode = progressive ? &TVEurgb60Hz480Prog : &TVMpal480IntDf;
+        break;
+
+    case CONF_VIDEO_NTSC:
+        r_rmode_reg = VI_NTSC;
+        r_rmode = progressive ? &TVNtsc480Prog : &TVNtsc480IntDf;
+        break;
+    }
+
+    switch (region)
+    {
+    case 'D':
+    case 'F':
+    case 'P':
+    case 'X':
+    case 'Y':
+        r_rmode_reg = PAL60 ? VI_EURGB60 : VI_PAL;
+        r_rmode = progressive ? &TVEurgb60Hz480Prog : (PAL60 ? &TVEurgb60Hz480IntDf : &TVPal528IntDf);
+        break;
+    case 'E':
+    case 'J':
+        r_rmode_reg = VI_NTSC;
+        r_rmode = progressive ? &TVNtsc480Prog : &TVNtsc480IntDf;
+    }
+
+    (*(volatile unsigned int *)0x800000cc) = r_rmode_reg;
+    rrc_invalidate_cache((void *)0x800000cc, 4);
+
+    if (r_rmode != 0)
+    {
+        VIDEO_Configure(r_rmode);
+    }
 }
