+++ source/loader/disc_loader.c
@@ -1,0 +1,127 @@
+/*
+    disc_loader.c - disc loading and identification
+
+    Copyright (C) 2025  Retro Rewind Team
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program.  If not, see <https://www.gnu.org/licenses/>.
+*/
+
+#include <string.h>
+
+#include "../res.h"
+#include "../prompt.h"
+#include "../di.h"
+#include "disc_loader.h"
+#include "../util.h"
+
+int rrc_disc_loader_locate_data_part(u32 *data_part_offset)
+{
+    int res;
+    struct rrc_di_part_group part_groups[4] __attribute__((aligned(32)));
+    res = rrc_di_unencrypted_read(&part_groups, sizeof(part_groups), RRC_DI_PART_GROUPS_OFFSET >> 2);
+    RRC_ASSERTEQ(res, RRC_DI_LIBDI_OK, "rrc_di_unencrypted_read for partition group");
+
+    struct rrc_di_part_info partitions[4] __attribute__((aligned(32)));
+
+    for (u32 i = 0; i < 4 && *data_part_offset == UINT32_MAX; i++)
+    {
+        if (part_groups[i].count == 0 && part_groups[i].offset == 0)
+        {
+            // No partitions in this group.
+            continue;
+        }
+
+        if (part_groups[i].count > 4)
+        {
+            RRC_FATAL("too many partitions in group %d (max: 4, got: %d)", i, part_groups[i].count);
+        }
+
+        res = rrc_di_unencrypted_read(&partitions, sizeof(partitions), part_groups[i].offset);
+        RRC_ASSERTEQ(res, RRC_DI_LIBDI_OK, "rrc_di_unencrypted_read for partition");
+        for (u32 j = 0; j < part_groups[i].count; j++)
+        {
+            if (partitions[j].type == RRC_DI_PART_TYPE_DATA)
+            {
+                *data_part_offset = partitions[j].offset;
+                break;
+            }
+        }
+    }
+
+    return 0;
+}
+
+int rrc_disc_loader_await_mkw(void *xfb, char *region)
+{
+    int res;
+    unsigned int status;
+
+check_cover_register:
+    res = rrc_di_get_low_cover_register(&status);
+    RRC_ASSERTEQ(res, RRC_DI_RET_OK, "rrc_di_getlowcoverregister");
+
+    // if status = 0 that means that a disk is inserted
+    if ((status & RRC_DI_DICVR_CVR) != 0)
+    {
+    missing_mkwii_alert:
+        char *lines[] = {
+            "Mario Kart Wii is not inserted!",
+            "",
+            "Please insert Mario Kart Wii into the console,",
+            "and select OK when done."};
+
+        enum rrc_prompt_result pres = rrc_prompt_ok_cancel(xfb, lines, 4);
+        RRC_ASSERT(pres != RRC_PROMPT_RESULT_ERROR, "failed to generate prompt");
+
+        if (pres == RRC_PROMPT_RESULT_OK)
+        {
+            goto check_cover_register;
+        }
+        else
+        {
+            return RRC_RES_SHUTDOWN_INTERRUPT;
+        }
+    }
+
+    rrc_dbg_printf("check disc");
+
+    /* we need to check we actually inserted mario kart wii */
+    struct rrc_di_disk_id did;
+    res = rrc_di_get_disk_id(&did);
+    /* likely drive wasnt spun up */
+    if (res != RRC_DI_LIBDI_EIO)
+    {
+        /* spin up the drive */
+        rrc_dbg_printf("failed to read disk_id: attempting drive reset\n");
+        RRC_ASSERTEQ(rrc_di_reset(), RRC_DI_LIBDI_OK, "rrc_di_reset");
+        res = rrc_di_get_disk_id(&did);
+        RRC_ASSERTEQ(res, RRC_DI_LIBDI_OK, "rrc_di_get_disk_id (could not initialise drive)");
+    }
+
+    /* this excludes region identifier */
+#define DISKID_MKW_ID "RMC"
+    if (memcmp(did.game_id, DISKID_MKW_ID, strlen(DISKID_MKW_ID)))
+        goto missing_mkwii_alert;
+
+    char gameId[16];
+    snprintf(
+        gameId, sizeof(gameId), "%c%c%c%cD%02x", did.game_id[0],
+        did.game_id[1], did.game_id[2], did.game_id[3], did.disc_ver);
+
+    rrc_dbg_printf("Game ID/Rev: %s\n", gameId);
+    memcpy((u32 *)0x80000000, &did, sizeof(did));
+    *region = did.game_id[3];
+
+    return RRC_RES_OK;
+}
